mission: step5 리펙토링
