+++ README.md
@@ -1,1 +1,22 @@
 # kotlin-racingcar
+
+## 기능 요구사항
+- [x] 횟수를 입력 받는다
+- [x] 자동차는 이름을 가진다
+- [x] 자동차는 전진할 수 있다
+- [x] 자동차는 움직이지 않을 수 있다
+- [x] 자동차는 위치를 가진다
+- [x] 자동차의 초기 위치는 1이다 ("-" 하나를 그리기 위해)
+- [x] 자동차는 무작위 숫자중 4이상일 경우 전진한다
+- [x] 자동차는 무작위 숫자중 3이하일 경우 움직이지 않는다
+- [x] 자동차의 이름은 ,(콤마)를 기준으로 나뉜다
+- [x] 자동차의 이름은 5자를 초과할 수 없다
+- [x] 자동차의 상태를 출력할 수 있다
+- [x] 자동차의 상태는 이름과 위치를 포함한다
+- [x] 자동차 게임이 완료되었을 경우 우승자를 출력한다
+- [x] 우승자는 한명 이상일 수 있다
+- [x] 우승자는 가장 많이 전진한 자동차이다
+- [x] UI 로직과 핵심 로직을 분리한다
+- [x] indent(인덴트, 들여쓰기) depth를 2를 넘지 않도록 구현한다
+- [x] 함수(또는 메소드)의 길이가 15라인을 넘어가지 않도록 구현한다
+- [x] 함수(또는 메소드)가 한 가지 일만 잘 하도록 구현한다
