+++ src/main/kotlin/first_week_carrace/Main.kt
@@ -1,5 +1,7 @@
 package first_week_carrace
 
+import first_week_carrace.controller.CarRaceGame
+
 fun main() {
     CarRaceGame().startGame()
 }
