+++ src/main/kotlin/first_week_carrace/model/ForwardDecider.kt
@@ -1,4 +1,4 @@
-package first_week_carrace
+package first_week_carrace.model
 
 fun interface ForwardDecider {
     fun execute(): Boolean
