+++ src/main/kotlin/first_week_carrace/utils/Constant.kt
@@ -1,0 +1,6 @@
+package first_week_carrace.utils
+
+const val MAX_NAME_LENGTH = 5
+const val RANDOM_START_RANGE = 0
+const val RANDOM_END_RANGE = 9
+const val MOVE_THRESHOLD = 4
