+++ src/main/kotlin/first_week_carrace/view/Viewer.kt
@@ -1,4 +1,4 @@
-package first_week_carrace
+package first_week_carrace.view
 
 class Viewer {
