+++ src/test/kotlin/first_week_carrace/CarRaceTest.kt
@@ -1,5 +1,6 @@
 package first_week_carrace
 
+import first_week_carrace.model.CarRace
 import org.assertj.core.api.Assertions.assertThat
 import org.assertj.core.api.Assertions.assertThatIllegalArgumentException
 import org.junit.jupiter.api.Test
